lab4 class line including states
Lab4) The class Line includes an attribute state = ’free’ or ’occupied’ (initialized as ’free’). This attribute shows if a connection is already occupying that line. All the lines within the path have to be ’free’.  If there are not any available path between the input and the output nodes of a connection, the resulting snr and latency have to be set to zero and ’None’, respectively.