Lab4 main
Lab4) we simulate few cases of signal propagation in our software abstraction of the network.  The main constructs the network defined by ’nodes.json’ and runs its method stream over 100 connections with signal power equal to one and the input and output nodes randomly chosen. It computes and plot the distribution of the best path in terms of latency or SNR.