updating stream metho to call calculate_bit_rate
Modify the stream() method of the Network class that has to call the calculate bit rate(path, strategy) once the path for the connection is given and using the transceiver attribute value of the first node in the path. If the path does not reach the minimum GSNR requirement for the specified transceiver strategy (zero bit rate case), the connection has to be rejected. Add the attribute bit rate to the Connection class that stores the assigned bit rate Rb.